models moved to separate package for clarity
