+++ src/main/java/com/stocktrading/dataloader1/domain/KafkaEventPublisher.java
@@ -1,6 +1,8 @@
 package com.stocktrading.dataloader1.domain;
 
 
+import com.stocktrading.dataloader1.domain.model.FinancialInstrumentPriceModel;
+
 import java.util.List;
 
 public interface KafkaEventPublisher {
