+++ src/main/java/com/stocktrading/dataloader1/domain/StockPriceReceivedEvent.java
@@ -1,5 +1,6 @@
 package com.stocktrading.dataloader1.domain;
 
+import com.stocktrading.dataloader1.domain.model.FinancialInstrumentPriceModel;
 import lombok.Getter;
 import org.springframework.context.ApplicationEvent;
