+++ src/main/java/com/stocktrading/dataloader1/domain/model/FinancialInstrumentModel.java
@@ -1,4 +1,4 @@
-package com.stocktrading.dataloader1.domain;
+package com.stocktrading.dataloader1.domain.model;
 
 import lombok.Builder;
