+++ src/main/java/com/stocktrading/dataloader1/remote/kafkaEventPublisher/StockDataEventPublisher.java
@@ -1,7 +1,7 @@
 package com.stocktrading.dataloader1.remote.kafkaEventPublisher;
 
 import com.stocktrading.dataloader1.domain.KafkaEventPublisher;
-import com.stocktrading.dataloader1.domain.FinancialInstrumentPriceModel;
+import com.stocktrading.dataloader1.domain.model.FinancialInstrumentPriceModel;
 import com.stocktrading.dataloader1.domain.StockPriceReceivedEvent;
 import lombok.AllArgsConstructor;
 import lombok.extern.log4j.Log4j2;
